1. Report error when setting deprecated attributes of Omega (O.engines, O.dt, ...) 2. Compile-time errors for undefined gl-templates 3. Render walls using GLUtils::Grid 4. Support compilation without VTK in gl/Renderer  (no Sparc) 5. Fix unit tests with yade.cld imports 6. Remove the DEM renderer from Renderer, use Gl1_DemFIeld instead 7. Remove the Sparc renderer, is in Gl1_SparcField now (not tested) 8. Handle imports of clumps from clDem 9. Handle comparison of clumps between clDem and yade

+++ pkg/dem/ContactLoop.cpp
@@ -143,8 +143,8 @@ void ContactLoop::run(){
 				const shared_ptr<Shape>& sh(particle->shape);
 				if(!sh) continue;
 				if(sh->nodes.size()!=1){
-					for(int i=0; i<sh->nodes.size(); i++){
-						if(sh->nodes[i]->getData<DemData>().flags&DemData::DOF_ALL!=DemData::DOF_ALL) LOG_WARN("Multinodal #"<<particle->id<<" has free DOFs, but force will not be applied; set ContactLoop.applyForces=False and use IntraForce(...) dispatcher instead.");
+					for(size_t i=0; i<sh->nodes.size(); i++){
+						if((sh->nodes[i]->getData<DemData>().flags&DemData::DOF_ALL)!=DemData::DOF_ALL) LOG_WARN("Multinodal #"<<particle->id<<" has free DOFs, but force will not be applied; set ContactLoop.applyForces=False and use IntraForce(...) dispatcher instead.");
 					}
 					continue;
 				}
